functions/add-url-pub/requirements.txt
google-cloud-pubsub==1.5.*
google-cloud-secret-manager==1.0.*
python-telegram-bot==12.7.*
